profile: Silence has_homebrew change on msys
Fixes #2.

I don't know why I haven't seen this before; the `profile` has been used on macOS and Linux previously without any issues like this. But, apparently the version of GNU `which` distributed with msys differs in its verbosity level when the provided command is not found in the path.