xsessionrc: added to repo
This one is important since by default, IntelliJ uses the current locale
when printing "git blame" dates. This will use MM/DD/YY format for me,
unless LC_CTIME is overriden. To be able to launch IntelliJ from the
Cinnamon Menu, it needs to be set as part of the X session, i.e. before
Cinnamon is started.