vscode: Update keybindings
It turned out that the config on my work desktop machine had gone out of
sync with this git repo (it wasn't a symlink but a regular file). I had
a really nice binding in my work machine that I wanted elsewhere, so I
overwrote these completely with my work bindings for now.