gitconfig: Add core.autocrlf, core.pager and change push.default
These matches the default on my work desktop computer. The driving force
for this was `core.pager` - git has an annoying idea of putting `git
show` output through the pager regardless of whether this actually makes
sense or not. The `less -F` flag (which comes for free with the
`diff-so-fancy` setup) overturns this, so that paging is avoided when
not needed, which gives a much more pleasant experience.