bash_aliases: Support 'main' branch name in grbm() function
This is a followup to 032fa5b01c315a04ff22cc546ad7dde9bfd293b6. When
working with an upstream repo where the default branch is named `main`
instead of `master`, this function would not work correctly. The
workaround is to add some extra checking here to determine the name of
the default branch.