bash_aliases: Add redshift aliases
I sometimes have to do late-night deployments (at 01:00 AM), and it's
useful to be able to shift the colors of the screen a bit while doing
so. `redshift` has an "automatic" mode as well, but the manual way of
enabling/disabling it as needed works fine for my use case.