bash_aliases: Add dark/light mode aliases
These play well with the changes added in
8421fcaaeb40e995d628a661e44d648ad9cbdbaf, to help make the cost of
switching between dark and light mode as low as possible.

Background: I tend to do a bit of
"hacking-in-the-evening-on-the-balcony" during the summer period. When
sitting outdoors with (setting, but still) sunlight, dark mode simply
makes the text too dark for me. Hence, I want to make it _easy_ to
switch back and forth between these themes.

With this change and the other commit mentioned, the following apps work
perfectly for me with seamless dark/light mode switching:

- Firefox (_especially_ nice with websites like GitHub which supports using "system theme" for dark/light mode)
- VS Code

Terminator also works, but not seamlessly; I have added a custom profile
for this. I might be able to make it support it seasmlessly as well, but
it's currently unclear how much work this is (to get it to use
a reasonable palette for ANSI-based colors).