(vscode) Do not specfiy workbench.colorTheme
This is intended to come from the `workbench.preferredLightColorTheme`
and `workbench.preferredDarkColorTheme` settings, depending on the
current dark/light mode system setting (provided by the GTK (?) theme).

However, the above doesn't seem to work flawlessly on my work/desktop
machine (`alladin`), whereas it worked fine on my work laptop
(`vostro`). I wonder if there is some package that needs to be updated
on this machine to make it work...