(firefox-install) Added script for installing Firefox from .tar.bz2
This is one of the things that I'd rather not have had to do, but...
reality leaves me with few options. The Debian package is simply often
older than I'd prefer, and I'm not (yet) willing to embrace the
Snap-based approach. Hence, a simple wrapper script which just automates
the process of initial install of Firefox will suffice for now.