(gitconfig) Remove `rebase.autosquash = true`
This reverts commit f0b47f0f1514666a6e5e3e8fd2551415b01e89ad and
d0fa80a800900f4b6941230819b2e23f1ef52aab.

It turned out that this breaks `git rebase`, in a fundamental way: it
will no longer output "Current branch master is up to date" when
performing a rebase, when this flag is active. This is a fundamental git
feature to me that I use to be able to determine if the rebase did
actually change anything or not, so I will have to live with manually
adding `--autosquash` when needed.