(bash_aliases) Add new alias
A convenience alias for converting bank statements ("kontoutdrag" in
Swedish) from their canonical PDF form to HTML format.

Ideally, I would like to have the statements in a programmatically
readable form, like CSV or JSON. Unfortunately, the CSV export available
from my bank does not seem to include all data, unlike the PDF
statements which have multiple lines for the message field. For now,
converting the PDF statements to HTML is one of the better ways to work
with these files, if you e.g. need to look for a particular transaction
in multiple years of data.